Bring in ColorUtils, take out TextraTypist dep.
ColorUtils is from TextraTypist and was copied almost in full; the color name description code is too ingrained to copy easily, so I just didn't bring it over. It's up to you if ColorUtils belongs in Stripe so other code of yours can use it, or just in the color picker library since that deals with colors primarily.